Separate utils module for the functions
